Move get_coordinates to MapService
